quantum_ml_time_series_project/src/self_reflection.py

import numpy as np
from qiskit import QuantumCircuit, Aer, transpile, execute

class QuantumSelfReflection:
    def __init__(self, num_qubits, memory_size=10, alpha=0.1, gamma=0.9):
        self.num_qubits = num_qubits
        self.memory_size = memory_size
        self.alpha = alpha  # Learning rate
        self.gamma = gamma  # Discount factor

        # Memory stores past quantum states and actions
        self.memory = []
        self.qc = QuantumCircuit(num_qubits)
        self.simulator = Aer.get_backend('qasm_simulator')

    def _quantum_encoding(self, state):
        """Encodes classical state into quantum state."""
        self.qc.reset(range(self.num_qubits))
        for i, bit in enumerate(state):
            if bit == 1:
                self.qc.x(i)
        return transpile(self.qc, self.simulator)

    def _quantum_state_to_classical(self):
        """Measure quantum state and return classical state."""
        self.qc.measure_all()
        job = execute(self.qc, self.simulator, shots=1)
        result = job.result().get_counts()
        measured_state = list(result.keys())[0]
        return int(measured_state, 2)

    def reflect(self, state, action):
        """Self-reflect on past quantum states and actions."""
        self.memory.append((state, action))
        if len(self.memory) > self.memory_size:
            self.memory.pop(0)  # Keep memory within limits

        # Reinforce learning based on self-reflection
        quantum_state = self._quantum_encoding(state)
        classical_state = self._quantum_state_to_classical()
        future_rewards = [self._estimate_future_reward(s, a) for s, a in self.memory]
        return np.mean(future_rewards)

    def _estimate_future_reward(self, state, action):
        """Estimate the future reward from a quantum state."""
        quantum_state = self._quantum_encoding(state)
        classical_state = self._quantum_state_to_classical()
        reward = np.random.rand()  # Placeholder for a more complex reward function
        return reward * self.gamma

    def reset_memory(self):
        """Reset the reflection memory."""
        self.memory = []

    def save_memory(self, path):
        """Save the reflection memory to a file."""
        np.save(path, self.memory)

    def load_memory(self, path):
        """Load the reflection memory from a file."""
        self.memory = np.load(path, allow_pickle=True).tolist()
